refactor: move work directory methods to separate class
